Completing exercices. added categories . +bonus exercices
