+++ app/src/main/java/com/ElOuedUniv/maktaba/domain/usecase/GetCategoriesUseCase.kt
@@ -8,6 +8,6 @@ class GetCategoriesUseCase(
     private val categoryRepository: CategoryRepository
 ) {
     operator fun invoke(): List<Category> {
-        return categoryRepository.getAllCategories()
+        return categoryRepository.getAllCategories().sortedBy { it.name }
     }
 }
